implemtation of column chart CarrierProfile
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C23311

+++ TOneV2/Code/TOneV2/TOne.Analytics.Data.SQL/BillingStatisticDataManager.cs
@@ -82,7 +82,7 @@ namespace TOne.Analytics.Data.SQL
             MonthTraffic instance = new MonthTraffic
             {
                 Month = reader["Month"] as string,
-                Durations = GetReaderValue<double>(reader, "Durations"),
+                Durations = GetReaderValue<decimal>(reader, "Durations"),
                 Amount = GetReaderValue<double>(reader, "Amount")
             };
